requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==39.0.2
paramiko==3.1.0
Pillow==9.4.0
pycparser==2.21
PyNaCl==1.5.0
